requirements.txt
# Core
pandas>=2.0
numpy>=1.24

# Plotting
matplotlib>=3.7
plotly>=5.18

# File watching / GUI
watchdog>=4.0

# Excel I/O
openpyxl>=3.1
xlsxwriter>=3.2
# If your plots tool reads DBF files, enable the next line:
# simpledbf>=0.2

# PLC
python-snap7>=1.1

# Jupyter (for Modbus notebook)
notebook>=7.0
jupyterlab>=4.0
ipykernel>=6.25
